Add client webpack configuration
There is a common base configuration and then the dev and production
configs layer onto that. The dev config is meant to be run with
webpack-dev-server which will give an html page to view everything. This
can be moved over to generate it for us, or it can be served from
express via a template.